+++ src/Packing/CartItem.php
@@ -28,9 +28,9 @@ class CartItem extends Item {
 
 		$s_product = wc_gzd_shipments_get_product( $this->get_product() );
 
-		$width  = empty( $s_product->get_shipping_width() ) ? 0 : wc_format_decimal( $s_product->get_shipping_width() );
-		$length = empty( $s_product->get_shipping_length() ) ? 0 : wc_format_decimal( $s_product->get_shipping_length() );
-		$depth  = empty( $s_product->get_shipping_height() ) ? 0 : wc_format_decimal( $s_product->get_shipping_height() );
+		$width  = empty( $s_product->get_shipping_width() ) ? 0 : (float) wc_format_decimal( $s_product->get_shipping_width() );
+		$length = empty( $s_product->get_shipping_length() ) ? 0 : (float) wc_format_decimal( $s_product->get_shipping_length() );
+		$depth  = empty( $s_product->get_shipping_height() ) ? 0 : (float) wc_format_decimal( $s_product->get_shipping_height() );
 
 		$this->dimensions = array(
 			'width'  => (int) wc_get_dimension( $width, 'mm' ),
@@ -38,7 +38,7 @@ class CartItem extends Item {
 			'depth'  => (int) wc_get_dimension( $depth, 'mm' ),
 		);
 
-		$weight        = empty( $this->get_product()->get_weight() ) ? 0 : wc_format_decimal( $this->get_product()->get_weight() );
+		$weight        = empty( $this->get_product()->get_weight() ) ? 0 : (float) wc_format_decimal( $this->get_product()->get_weight() );
 		$quantity      = (int) ceil( (float) $item['quantity'] );
 		$line_total    = (int) wc_add_number_precision( $this->item['line_total'] );
 		$line_subtotal = (int) wc_add_number_precision( $this->item['line_subtotal'] );
